add `Module` superclass for module inheritance, refactor `sItem -> SItem` and inherit `Module` interface
