--- src/config/base.json
@@ -1,13 +1,0 @@
-{
-  "ENABLED": false,
-
-  "TIME_TO_DEPLOY": 20,
-
-  "INSTANT_HIDEOUT_CONSTRUCTION": false,
-  "CONSTRUCTION_TIME_MULT": 1.0,
-  "INSTANT_HIDEOUT_PRODUCTION": false,
-  "PRODUCTION_TIME_MULT": 1.0,
-
-  "DISABLE_FLEA": false,
-  "FLEA_LEVEL": 15
-}
