--- src/config/game.json
@@ -1,20 +1,0 @@
-{
-  "SCAV_COOLDOWN": 1200,
-
-  "SURVIVED_EXP_MULT": 1.0,
-  "KILLED_EXP_MULT": 1.0,
-
-  "NOTE (I)": "the following represent maximum player health values",
-
-  "HEAD_HP": 35,
-  "CHEST_HP": 85,
-  "STOMACH_HP": 70,
-  "LARM_HP": 60,
-  "RARM_HP": 60,
-  "LLEG_HP": 65,
-  "RLEG_HP": 65,
-
-  "SKILL FATIGUE": true,
-
-  "EXTRACT_MASTER": false
-}
