--- src/config/misc.json
@@ -1,11 +1,0 @@
-{
-  "AIRDROPS": "the following represent chance percentage",
-
-  "CUSTOMS": 0.0,
-  "WOODS": 0.0,
-  "LIGHTHOUSE": 0.0,
-  "SHORELINE": 0.0,
-  "STREETS": 0.0,
-  "RESERVE": 0.0,
-  "INTERCHANGE": 0.0
-}
