--- src/config/npc.json
@@ -1,53 +1,0 @@
-{
-  "USEC_BEAR_RATIO": 0,
-  "FRIENDLY_PMC_PERC": 0,
-  "MAX_BACKPACK_VALUE": 0,
-  "MAX_POCKET_VALUE": 0,
-  "MAX_VEST_VALUE": 0,
-
-  "NOTE (I)": "the following take min/max integers",
-
-  "SCAV_TO_PMC": [ 0, 0 ],
-  "CURSED_TO_PMC": [ 0, 0 ],
-  "RAIDER_TO_PMC": [ 0, 0 ],
-  "ROGUE_TO_PMC": [ 0, 0 ],
-
-  "BOT_WAVES": true,
-
-  "SCAV_HP_MULT": 1.0,
-  "SPEC_HP_MULT": 1.0,
-  "BOSS_HP_MULT": 1.0,
-  "FOLL_HP_MULT": 1.0,
-
-  "NOTE (II)": "the following take in chance/followers/location, where applicable",
-
-  "RESHALA": [ 35, "4" ],
-  "CUSTOMS_GOONS": [ 35 ],
-  "CUSTOMS_CULTIST": [ 15, "4" ],
-
-  "SHTURMAN": [ 35, "2" ],
-  "WOODS_GOONS": [ 35 ],
-  "WOODS_CULTIST": [ 15, "4", "ZoneMiniHouse,ZoneBrokenVill" ],
-
-  "INTERCHANGE_KILLA": [ 30 ],
-
-  "FACTORY_DAY_TAGILLA": [ 30 ],
-  "FACTORY_NIGHT_TAGILLA": [ 23 ],
-  "FACTORY_NIGHT_CULTIST": [ 10, "2" ],
-
-  "ZRYACHIY": [ 100, "2" ],
-  "LIGHTHOUSE_GOONS": [ 35, "Zone_TreatmentContainers,Zone_Chalet" ],
-
-  "SANITAR": [ 35, "2,3", "ZoneGreenHouses,ZoneSanatorium1,ZoneGreenHouses,ZoneSanatorium2,ZonePort" ],
-  "SHORELINE_GOONS": [],
-  "SHORELINE_CULTIST": [ 12, "3" ],
-
-  "KABAN": [ 35, 6 ],
-  "STREETS_KILLA": [ 5 ],
-
-  "NOTE (III)": "glukhar uses: chance/scouts/assault/security",
-
-  "GLUKHAR": [ 35, "2", "2", "2" ]
-
-
-}
