+++ src/helpers/ExitGenerator.ts
