+++ src/helpers/ExitHandler.ts
