+++ src/mod.ts
@@ -1,24 +1,21 @@
 import { DependencyContainer } from "tsyringe";
 import { IPostDBLoadMod } from "@spt-aki/models/external/IPostDBLoadMod";
-import Index from "./index";
-
-const index = new Index();
-
 import * as config from "../config/data.json";
-import sBase from "./overrides/sBase";
-import sAirdrop from "./overrides/sAirdrop";
-import sBoss from "./overrides/sBoss";
-import sBot from "./overrides/sBot";
-import sHideout from "./overrides/sHideout";
-import sPmc from "./overrides/sPmc";
-import ExitHandler from "./modules/ExitHandler";
+
+import sBase from "./modules/sBase";
+import sAirdrop from "./modules/ShAirdrop";
+import sBoss from "./modules/sBoss";
+import sBot from "./modules/ShBot";
+import sHideout from "./modules/ShHideout";
+import sPmc from "./modules/ShPmc";
+import ExitHandler from "./helpers/ExitHandler";
 
 export default class Mod implements IPostDBLoadMod
 {
     public postDBLoad(container: DependencyContainer): void
     {
         const modules: readonly
-        new ExitHandler(config);
+        //new ExitHandler(config);
     }
 }
