+++ src/modules/ShAirdrop.ts
