+++ src/modules/ShBot.ts
