+++ src/modules/ShHideout.ts
